Phase 4.25.1: Fix build error — num -> double for HSLColor hue
Build failed because (hash % 80) returns num (int), but
HSLColor.fromAHSL() requires double for the hue parameter.

Fix: cast via .toDouble() before passing.

Error was:
  lib/app/ui/screens/genres_tags_collections_page.dart:739:16: Error:
  The argument type 'num' can't be assigned to the parameter type 'double'.
        hue < 30 ? hue + 350 : hue,
                 ^

+++ lib/app/ui/screens/genres_tags_collections_page.dart
@@ -733,7 +733,7 @@ class _CollectionCardState extends State<_CollectionCard> {
     }
     // Restrict hue to warm red-orange-pink range (0–50, 330–360) so
     // avatars stay on-brand with the Netflix-red palette.
-    final hue = (hash % 80);
+    final hue = (hash % 80).toDouble();
     return HSLColor.fromAHSL(
       1.0,
       hue < 30 ? hue + 350 : hue,
